Zoom out when on small screens :)
This does absolutely nothing on firefox, but works great in safari and chrome.

There are tricks to make zooming happen on firefox, but it involves stuff like this

```
        html {
          -moz-transform: scale(0.7);
          -moz-transform-origin: 0 0;
        }
        @-moz-document url-prefix() {
          html {
            width: 142.857%;
          }
        }
```

amd after that some things still have a very hard time staying aligned with the screen edges (such as the sidebar). So I've left that as a future exercise.