Pseudo-code some basic tests and leverage useTheme hook
The test suite cannot currently be run, as the react-syntax-highlighter
appears to be using ESM, which are not supported by Jest out-of-the-box.

This should only require a simple Babel loader to remedy, but I want to
make sure first that I have not misconfigured something, as configuring
Babel is likely a notable scope increase for this relatively small
issue.