ADR: Core entities in the catalog
This is capturing the catalog part of the Backstage System Model
RFC (#390). I'm planning to follow this up with a second ADR on
how to express System, Application and Domain in catalog labels.

Not sure if it helps to have the example yamls inline here. It
helps me to visualize how it'd work, but it could also get outdated
quickly.