Refactor RadarPlot & Best effort with PropTypes
A global search looks like no other plugins use PropTypes. I'm about to
remove them in the next commit because shapes don't seem to cooperate
with TypeScript, making the PropTypes either so vague that they're
mostly useless or specific enough to fail the build.

With that said, I wanted to get this version committed in case I need to
roll back to here. I wouldn't want to redo this work.