Align types
I'm not confident in this change. I'm very unclear about the reasoning
behind the type definitions in api.ts as they don't reflect the shape of
the values that I see when I print `data.entries` in the console.

The value of data.entries reflects a fully-populated Entry[], not a
RadarEntry[].

As such, I'm also not sure about the sample data. I would be more
hesitant to updated it, but the fact that the rest of the components
seem to be expecting that data makes it seem fairly clear that it must
be required to be in the updated shape.