Issue 661 Tech Radar Typescript refactor (#1080)
* Refactor RadarFooter

* Refactor RadarLegend Pt 1

* Refactor RadarLegend Pt 2

Seeing some errors that I don't fully understand around the proptypes
assertions seemingly being treated as null | undefined values.

It looks like I'll need to type Segment, which is not explicitly
in-scope for this issue.

* Fix segments regression

I incorrectly added type safety to the original implementation, causing
the segments to no longer be populated here. This change fixes that
regression while retaining the type safety.

This is all necessary because qidx and ridx are apparently both
nullable.

* Resolve remaining RadarLegend warnings

* Refactor RadarEntry

* Forcibly skip hooks to revert migration prettify

* Refactor RadarGrid

* Refactor RadarPlot & Best effort with PropTypes

A global search looks like no other plugins use PropTypes. I'm about to
remove them in the next commit because shapes don't seem to cooperate
with TypeScript, making the PropTypes either so vague that they're
mostly useless or specific enough to fail the build.

With that said, I wanted to get this version committed in case I need to
roll back to here. I wouldn't want to redo this work.

* Remove PropTypes from refactored components

The build now passes.

* Refactor RadarBubble

* Refactor Radar

* Move Radar functions out of component file

I could see these being split into separate files, but I think this is
enough for now.

* Small function signature change

* Align types

I'm not confident in this change. I'm very unclear about the reasoning
behind the type definitions in api.ts as they don't reflect the shape of
the values that I see when I print `data.entries` in the console.

The value of data.entries reflects a fully-populated Entry[], not a
RadarEntry[].

As such, I'm also not sure about the sample data. I would be more
hesitant to updated it, but the fact that the rest of the components
seem to be expecting that data makes it seem fairly clear that it must
be required to be in the updated shape.