Force code to use postgres via knex
Here I am attempting to find all points in the codebase which
instanciate a database so I can force them to use postgres. This is just
for development purposes so I can be sure I am using PG. It is kinda
confusing because, for example, the backend package has a `knexfile.js`
which is completely ignored.