The backend listens to all interfaces by default
Previously,  if you did not specify a host,  it would only
listen to localhost,  so it would not accept connections from the network.

If you use the default generated app,  it is configured to listen to 0.0.0.0
which only listens to IPv4 interfaces.

This changes the default behavior, just listen to all interfaces.