Update the system model to focus more on the core entities.
Now that we have a way of labeling and annotating catalogue
entities, we can focus the system model more on the entities
themselves. I believe that also makes more sense as a model that
is shared across many companies - it's easier to standardize on
the core entities, and leave some room to implement further
abstraction or categories to individual implementations.