Update Refresh Token, if provided by the server in the refresh grant
The OAuth2 specifications specify that during the refresh grant, the
server may provide an alternate refresh Token along with the access
Token when performing a token refresh.
The oauth2 provider in the `auth-backend` ignores the newer refresh
token during the refresh grant.
This commit will update the refresh token cookie at the end of the
call to the /auth/oauth2/refresh endpoint if the server provides a new
refresh token
grant a