Conditionally add Sentry widget to component page
This adds the Sentry widget back to the component page (from which it
was removed in #2246) if the component declares its own Sentry project
ID in the annotation named backstage.io/sentry-project-id. The widget
is hidden otherwise.

Previous to #2246, the widget was just displaying an empty list because
it was trying to load data for a hardcoded project name of
sample-sentry-project-id.

Fixes #2257