Assorted scaffolder form help text improvements (#2467)
* Prompt the user for a description in templates

The text "Description of the component" was a little redundant since the
label already says "Description". This description text hints to the
user why they should bother adding an informative description.

* Inform user about lowercase, url-safe constraint

Create React App scaffolding fails when a name with spaces or other
non-url-safe chars are provided. Instead of making the user wait for a
failure, Backstage should tell them about this constraint up front.