feat (techdocs): Enable techdocs without docker-in-docker
When the backend is deployed onto a Docker container, the techdocs volume mount logic
does not seem to work. A workaround until a more permanent solution is found is here:
https://github.com/spotify/backstage/pull/2438

Updated our fork to just include this update, but we should migrate to the more permanent
solution when it's released.