techdocs: Remove spinners
Spinners cause unnecessary UX distraction.
Case 1 (when docs are built and are to be served): Spinners appear for a split second before the name of site shows up. This unnecessarily distracts eyes because spinners increase the size of the Header. A dot (.) would do fine. Definitely more can be done.
Case 2 (when docs are being generated): There is already a linear progress bar (which is recommended in Storybook). Spinners are not good. 🤷