Improve SEO of microsite title (#3608)
Many people will be searching for "service catalog" and "developer platform" when they are looking for something like Backstage. They may not have ever heard of Backstage before so they will not know to search for it.

A title which includes the keywords "Service Catalog" and "developer platform" will rank better for those queries. This increases the change of Backstage being found by people who could make good use of it.

[Standard advice](https://moz.com/learn/seo/title-tag) is to keep page titles under 60 chars. The proposed title is 49 chars.