Implement readTree for GitlabReader (#3547)
* This change adds an implementation for the readTree method in GitlabReader.
The implementation is similar to the readTree method in GithubReader
with the exception of the 'path' variable used for subpaths in the
archive. To facilitate this, GitlabIntegrationConfig now has an apiUrl
field which was missing for Gitlab but was present for Bitbucket & GH.

* As part of this change, there are also new tests added to GitlabReader
to mirror what has been done with the GH reader. For now the archive
format chosen is 'zip', follow-up action includes having a look at
whether tar.gz is preferable.

Signed-off-by: Matei David <madavid@expediagroup.com>