Experiment with a processor that finds all catalog-info.yaml files in a repository
We have bigger mono repositories with a lot of components and teams are wondering if they really have to register every catalog-info.yaml file in Backstage. We thought a little bit about it and build a small processor that automatically registers all catalog-info.yaml files in a repository. This is not only done once, but on every refresh which means new catalog-info.yaml files are automatically deteceted and registered.

As this solution has some drawbacks, I would like to get some feedback before finishing this PR. There are also some implementation questions I have.

I think the biggest drawback is that we have to download the full repository archive every time just to find the catalog-info.yaml, which can be slow. We don't even read the contents but are only interested in the directory tree! I tried this with one of our bigger ones and it takes more than a minute.

I left some additional comments in the code, please have a look.