Fix issue where radar entry dialog with external url
If an entry has both a radar description and an external url, when
a user clicks on the blip the description dialog would briefly
appear before they are redirected to the specified url.

Expected behaviour is not to automatically redirect the user so that
they can read the dialog and choose to follow the link for more details.

Signed-off-by: Niall McCullagh <mccullagh_niall@hotmail.com>
Signed-off-by: Niall McCullagh <niallmccullagh@users.noreply.github.com>