Update the default CODEOWNERS to point to a new `reviewers` team
THIS IS A PROPOSAL.

Sets the fallback reviewers for pull requests to be a new `reviewers` team, which is a superset of `maintainers`.

The Backstage project sees a very large amount of activity. The project configuration currently has the `maintainers` team set as the gatekeepers of all pull requests, no matter how small or large. This puts pressure on the individual maintainers and can lead to longer lead times than desired.

As Spotify staffs up, and as external contributors move closer to the core of development, I propose that there needs to exist an intermediate step between "general contributor" and "full maintainer".

I propose that the set of people that shall have the rights to approve/merge pull requests is larger than just the maintainers. This would let us reduce pressure on individual maintainers and to expand the team at Spotify in an effective way. We could more easily perform goalie (rotating designated support person) duties with an extended team without using up critical maintainer voting slots.

It is important to note that the reviewers are not expected to behave as maintainers. Issues that are contested, or changes that have significant architectural effects on the code base, shall still be decided upon by maintainers. There is a certain amount of trust involved. This pull request puts the reviewers team at the root of the entire code base, but future tweaks may put the maintainers team as an owner of certain sub-directories if desired, to limit the reviewers' area of influence.

Signed-off-by: Fredrik Adelöw <freben@gmail.com>