Only show relevant tabs.
For organizations that have a lot of plugins, some teams may want to utilize plugins that others find to be a distraction on their pages.

The CI/CD switcher is a good pattern, but it still forces all entities of a type to have a tab that implements each category the entire organization has agreed upon.

With this change, the default EntityPage will only show the tabs that are applicable for the Entity based on its annotations, providing more control to implementing teams on how they want their Entity to look.

Signed-off-by: Andrew Thal <andrewthal88@gmail.com>