Allow backend builds without emitting types
The `backend:build` command by default, sets `Output.types` when calling `buildPackage`.
For backend Node JS builds, especially when building for production, types are superflous and the
code is transpiled to javascript and types are erased.
Despite types being not used, this lack of flexibility makes it necessary to emit types, that adds
a lot of accidental complexity.
Similar to the build command, optionally disable emitting types if a `--no-types` flag is passed

Signed-off-by: Gowind <petrovgovind@gmail.com>