Add `catalog.processors.microsoftGraphOrg.ignoreGroupIds`
Now that the new group page in the explore plugin shows the available groups nicely, we have to problem that we have a lot of garbage groups that we don't want to show. While we already have the `groupFilter` option, we can't build a filter that allows to exclude them. Even a filter that just excludes them by id is not implemented by the graph API, even though the syntax would support it.

Signed-off-by: Oliver Sand <oliver.sand@sda-se.com>