Set techdocs admonition font size to 1rem
Currently admonitions in TechDocs are rendered with font size 0.64rem which is much smaller than the rest of the text. This makes admontions hard to read and further "undo"s the attention they are here to provide.

This change adds a style overide, like we do with tables that sets the size to 1rem for bot collapsable and expanded admonitions.

Some trailing whitespace in the CSS styles is removed as well.

Signed-off-by: Crevil <bjoern.soerensen@gmail.com>