Allow configurable number of taskworkers
Right now we are limited to one taskworker in the scaffolder backend
which means you can only scaffold one thing at a time. This is a poor
user experience for larger organizations where multiple users maybe
scaffolding at the same time.

This adds an optional configuration option to increase the number
of taskworkers via the router options if you like, but defaults to
1 if not set.

Signed-off-by: jrusso1020 <jrusso@brex.com>