Describe publish:github scaffolder action fields
Currently the fields of the publish:github action uses some undocumented
conventions in the repoUrl and access input fields.

This change adds a description to the fields with examples of what to input. The
collaborators description is also expanded a bit to make it more clear that
these are additional compared to access and owner.

Signed-off-by: Crevil <bjoern.soerensen@gmail.com>