api-extractor: check whether folder has a package.json
Often when switching between branches, I have empty folders in my working copy and Git doesn't track folders. Another case is the now deleted `core` package which remains as an empty folder locally (due to the node_modules folder still being there).
Sadly, api-extractor crashes on empty folder, so this introduces a check that the folder is a package by checking for the package.json file.

Signed-off-by: Oliver Sand <oliver.sand@sda-se.com>