Codeowners processor should specify kind for user entities.
The codeowners processor extracts the username of the primary
owner and uses this as the owner field. Given the kind isn't
specified this is assumed to be a group and so the link to
owner in the about card doesn't work. This change specifies
the a kind where the entity is a user. e.g:
`@iain-b` -> `user:iain-b`

Signed-off-by: Iain Billett <iain@roadie.io>