add ability to specify an external id for assumed role for k8
From the AWS documentation.

A cross-account role is usually set up to trust everyone in an account.
Therefore, the administrator of the trusting account might send an external
ID to the administrator of the trusted account. That way, only someone with
the ID can assume the role, rather than everyone in the account.

Signed-off-by: Brian Fletcher <brian@roadie.io>