ability to discover catalog-infos from default branch
Previously the github-discovery supported only repos from a specific
branch. This change cause the discovery to work in two modes. It continues
to work as it did with a wildcard for repositories on a specified branch.

The new mode added causes the discovery processor to use the default branch
and looks for a file called catalog-info.yaml.

Signed-off-by: Brian Fletcher <brian@roadie.io>