feat(scaffolder-backend): Yeoman Generator module
Added a new Scaffolder Backend Yeoman module to allow for integration with Yeoman Generators. At the moment the only action added is 'run' to execute installed Yeoman Generator.

The Yeoman Generator can be either added as a dependency in `package.json` of a service that uses this plugin or installed in an environment using `npm install -g https://github.com/org/generator-name.git`

Tests for `yeomanRun.ts` were not added because 'yeoman'environment' module is exported in a way that does not allow partial mocking while '@types/yeoman-environment' does not support Yeoman 3.x, yet.

I also tested it E2E.

Signed-off-by: @pawelmitka <pawel.mitka@brainly.com>