fix(SignInPage): Ensure click handler and login popup occur in same tick
The current SignIn implementation is broken on Safari iOS by default, due to the combination of useEffect and the extra asynchronous login check that occurs before the authentication routine that causes the popup (instantPopup: true). This change refactors the internals of SignIn page to avoid useEffect, which was arguably confusing for implementors, and also avoid overloading the use of the poorly scoped autoSignIn variable.

Signed-off-by: Chris Carpita <ccarpita@butterflynetinc.com>