validate path directory, and prevent target from being overwritten
Validation now occurs to ensure that the target path exists and is a directory.
Additionally, instead of moving the temporary directory to the target path,
files are now recursively copied, and then cleaned up. This was to support app
creation in the current directory, without clobbering existing files (for
example .git/)

Instead of overwriting the target `path` directory while using the
move method,

Signed-off-by: Colton Padden <colton.padden@fastmail.com>