Template directly to specified path otherwise use temp dir
When the `--path` argument is specified, then perform the create-app templating
directly to that path. Otherwise, template to a temporary directory, and move
the files to a directory based on the project name.

Removed the `cleanUp` method in favor of removing temporary files directly in
the `moveApp` method in a `finally` block.

Reverted `moveApp` method to use the move operation instead ofe recursively
copying files.

Signed-off-by: Colton Padden <colton.padden@fastmail.com>