Export components so developers can customize the scaffolder page
This is the first step to allow developers to customize the scaffolder
page. With this work all the scaffolder page is now available to
backstage developers and they can customize the current scaffolder
page how they like similar to what is available for the catalog page
and other plugin pages.

Future work that has been discussed is possibly allowing for SwimLane
components that will allow developers to surface specific templates
or organize templates in specific ways

Signed-off-by: jrusso1020 <jrusso@brex.com>