plugins/scaffolder-backend: Fix a bug where only file mode 775 is considered executable.
Due to the default umask 002 on Ubuntu, the executable files checked out by Git
have file mode 100775 (33277), which by mistake, was assumed to be the only possible file mode
produced by Git.  This bug was introduced in https://github.com/backstage/backstage/pull/7738.

There are other possible file modes, such as 100700, 100755, which
should be considered executable.  This behavior is consistent with how
Git behaves.  Git only tracks whether a file is executable.

This pull request considers a file to be executable as long as there is
an execute bit in owner, group, or everyone.

Signed-off-by: Kenneth Feng <kenneth.feng@onepeloton.com>