authz: restrict type of PermissionCondition params
We need a guarantee that we'll be able to spread permission condition
params when applying the condition via the `apply` or `toQuery` methods
on the corresponding rule. As such, we need to enforce that the TParams
type parameter is an array.

This commit also supplies a default value of unknown[] for this type
param to make it a bit more convenient to use the type in situations
where the type of the params is not important.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>