permission-backend: remove standaloneServer, run.ts and AllowAllPermissionPolicy
The run.ts / standaloneServer.ts stuff is a hangover and expected to
be removed or heavily adjusted in the future, and since it's intended
to run from the command line, there's no neat way to pass in a
permission policy. As such it's not going to be useful in its current
form, so we're removing it for now.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>