adds the ability to configure the gh creds provider
This change provides an ability to pass a custom implementation of
the credentials provider via the backend package.

The motivation for this change is to make the private key of the
GitHub app more secure by allowing the backstage deployer to make
decisions about how the GitHub credentials are protected.

The GitHub app key gives great power to the holder and as such is
very sensitive. e.g. if you are using the scaffolder, you probably
have requested administration write permissions on your installations
organizations. The key needs to be protected well.

This change creates a new interface called IGitHubCredenialsProvider
and a GitHubCredentialsProviderFactory. Processors, URL readers etc
have been switched to uses the factory.

The factory can optionally be setup with a new provider as follows:

```
GithubCredentialsProviderFactory.setProvider(MyGithubCredentialsProvider);
```

This should be done before the plugins are configured in the backend
application. The default provider is the existing GithubCredentialsProvider.

Another proposal has been made to implement a dependency injection container.
The DI container approach would be better approach, however it requires more
discussion and design. This change brings us a step closer, because the DI
solution will require the new interface in any case.

Signed-off-by: Brian Fletcher <brian@roadie.io>