feat(integration): add support for a Bitbucket baseUrl config
With the addition of the Bitbucket BaseUrl configuration property, users can now explicitly define the baseUrl for their organization server. With that provided, processors can now strictly evaluate target urls instead of searching for the index of a potentially arbitrary term, which is now relegated as a backup solution.

If not specified, the config will be automatically constructed using the `host` config property. With this change, the config reader will also enforce the formatting of the `baseUrl` and `apiBaseUrl` properties to ensure that they are proper URL's.

Resolves #8771

Signed-off-by: Dhruval Darji <dhruvaldarji@gmail.com>