catalog: update DefaultCatalogPage test
This test appears to be checking for a regression in which the filter
reset from starred to owned after starring an entity. As a precursor,
it tried to select all three possible filters, but now it's no longer
possible to select filters that don't have any matching entities.

Now that the behaviour of menu items when there are no items in that
filter has changed, this commit updates the test to confirm that the
menu item starts disabled, and then enables when an entity is starred.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>