consolidate around a single aws-sdk version
The aws-sdk library is one of the largest depdencies of backstage. If there
are multiple versions in the picture, it can contribute to a larger image size.

This change brings the aws-sdk versions into line so that they are all the
same. While I was setting a version, I picked the most recent. Hopefully
this will be good enough.

Signed-off-by: Brian Fletcher <brian@roadie.io>