[wip] support 'splunk.com/on-call-routing-key' annotation
Extends the Splunk On Call plugin Entity card to infer the
corresponding team names from a `splunk.com/on-call-routing-key`
annotation.

This seeks to enable the use of the Splunk On Call plugin on
entity pages for those orgs who associate a component with a
routing key rather than a team name.

Signed-off-by: Mike Ball <mikedball@gmail.com>