Add techdocs-core plugin to techdocs builder automatically
Currently users need to modify their mkdocs files to container techdocs-core plugin manually when they want to expose techdocs in Backstage. To have a standardized set of extensions without the need to modify existing files (and possibly pollute alternative/additional mkdocs pipelines) this should be added automatically.

This PR adds a helper function to modify the mkdocs file to contain this plugin if it doesn't exist in the config file.

Co-authored-by: @emmaindal

Signed-off-by: Jussi Hallila <jussi@hallila.com>