Add a new interface: DocsBuildStrategy
This adds a new interface called DocsBuildStrategy. This strategy allows
for different answers to the question: should the TechDocs backend
perform a build of the given entity?

The default implementation replicates the existing functionality, in
that if the techdocs.builder config value is set to local, then we do
trigger builds, and if the value is set to anything else, we don't
trigger builds. However the strategy has access to the entity, and so
more complex strategies are possible.

This strategy is added as an optional parameter on the RouterOptions
types, and defaults to the DefaultDocsBuildStrategy if unspecified,
allowing for backwards compatibility.

TODO: rename some of the config options, and reword errors to reflect
the new interface

Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>