Fix header nesting in software-templates docs
Most pages in the 'Software Templates' docs seemed to have one level too
much nesting.  This fixes that.  Without the correct nesting, the table
of contents on the right side was also missing and made reading the
longer pages (e.g. writing-templates) more difficult to read.

What's odd is that when rendering with `mkdocs` locally the nesting
seems to be okay and the table of content renders.

Pages affected live:

- https://backstage.io/docs/features/software-templates/software-templates-index
- https://backstage.io/docs/features/software-templates/configuration
- https://backstage.io/docs/features/software-templates/writing-templates
- https://backstage.io/docs/features/software-templates/builtin-actions
- https://backstage.io/docs/features/software-templates/writing-custom-actions

https://backstage.io/docs/features/software-templates/adding-templates
is okay because it has no subheadings.

And these two are okay because they have correct nesting starting at
level two (`##`):

- https://backstage.io/docs/features/software-templates/writing-custom-field-extensions
- https://backstage.io/docs/features/software-templates/migrating-from-v1beta2-to-v1beta3

Signed-off-by: Luna Stadler <luc@spreadshirt.net>