catalog-common: remove resourceType from catalogEntityCreatePermission
The resourceType on permissions refers to the resource whose ref is
expected to be passed along with the permission during authorization.
This allows the permission-backend to make the decision based on
characteristics of the resource.

Since the entity being created by definition doesn't yet exist, it's
not correct for this permission to include a resourceType.

Signed-off-by: Mike Lewis <mtlewis@users.noreply.github.com>