search: throw error for unexpected non-resource permissions when authorizing result-by-result
We only reach this point in the code for types where the initial
authorization returned CONDITIONAL -- ALLOWs return early immediately
above, and types where the decision was DENY get filtered out entirely
when querying.

This means the call to isResourcePermission here is mostly about
narrowing the types, since the only way for it to be false is if the
PermissionPolicy returned a CONDITIONAL decision for a non-resource
permission, which can't happen - it would throw an error during
validation in the permission-backend.

Signed-off-by: MT Lewis <mtlewis@users.noreply.github.com>