Fix broken querySelectors on techdocs
This change fixes the broken pattern for selecting DOM elements to scroll to.

Previously it had error messages like so:
```
Uncaught DOMException: Failed to execute 'querySelector' on 'Element': '#fn:someidtag' is not a valid selector.
```

This did not bode well with the techdocs integration with footnotes.

The query has now changed to include something that is expected by the DOM queries.

Signed-off-by: Nicolas Arnold <nic@roadie.io>