feat: split `BitbucketClient` into `BitbucketCloudClient`, `BitbucketServerClient`
Bitbucket Cloud and Bitbucket Server (fka. Stash) are different product lines at Atlassian
and offer different APIs.

This was also reflected within the client which contained a mix of endpoints for both.

This is a tiny transparent step towards an easier maintainable setup as proposed at #9923.
The change is transparent for users of the processor, contained within the internals of it.

Besides the maintainability as of #9923, this can also enable future enhancements like
handling of rate limits, etc.

Some API interactions are not part of the client but are included in other packages like
`packages/integration` and then e.g. used at the `UrlReader` implementation.
This imposes some challenges towards topics like rate limit handling, etc.
Due to this, it may even make sense to externalize the clients in the future
(or alternative options) and ensure all interactions go through it.

As part of creating entity providers for such processors (#10183), this might be a good candidate
to create separate entity providers instead. It would not impact current users of the processor
(more than one entity provider).

Relates-to: #9923
Relates-to: #10183
Signed-off-by: Patrick Jungermann <Patrick.Jungermann@gmail.com>