Fix a path mismatch between techdocs embedded app and serve command
This fixes an issue introduced in
b44692890b1c6ca1df55313612a1d28e3c61ea8a.

An override of the path to the techdocs API was removed, without the
corresponding change in the Techdocs CLI. This resulted in traffic for
the techdocs API having an extra `techdocs` in the path, causing the
mkdocs server to 404.

Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>