Don't set the background color on Avatars with pictures.
When an Avatar component has a picture, the image may have a transparent
background. In that case, we don't know whether the generated background
color will work well with the image, or will clash with the image colors.
It's better to not generate a background color so the Avatar image will at
least render with predictable colors.

This fixes #11094.

Signed-off-by: James Peach <jpeach@cloudflare.com>