Don't swallow indexing failures
When an error occurs in the indexer pipeline, it logs it out, and then
drops it on the floor. This is problematic because externally, from
the task scheduler's perspective, it appears to have succeeded, which
is not true.

As a result, these failures are invisible to tooling which introspects
over tasks.

This makes the index task promise resolve when the pipeline finishes as
before, but it rejects if the pipeline failed.

Signed-off-by: Charles Lowell <cowboyd@frontside.com>