move register new component into context menu
We noticed that less than 1% of our users are creating new templates.
Thus I don't think this button should use such prime real state

We also performed a small usability test and we observed that
users expect the search box to be in this area.

So I would like to move the "Register New Component" button inside
the context menu, and call it "Add Template" instead.
This will also clear the space for the search box.

Signed-off-by: Paul Pacheco <paul.pacheco@aa.com>