Explain how to enable techdocs cache
For dummies like me, it took me a while to figure out that even if you
have configured the backend cache, the TechDocs cache is not enabled
until you set a value for the TTL. My interpretation was that it was
just configured by default, and the fact that a default value for ttl
seems to appear here points to that as well.

This change hopes to make this explicit, to avoid others spinning their
wheels like I have.

Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>