Add config hot reloading to proxy-backend
When working with the proxy configuration or plugins facilitating the proxy
package the workflow is currently somewhat cumbersom as changes to the proxy
configuration requires a full restart of the backend server.

Use cases where the proxy configuration is updated is when testing out new
routes for plugins to use and updating header configuration like authorization
tokens.

This change updates the proxy-backend to subscribe to configuration changes and
update the router when changes to the proxy is made.

Signed-off-by: Crevil <bjoern.soerensen@gmail.com>