Use Azure Identity to Authenticate with Microsoft Graph API
Replaced MSAL with Azure Identity to support additional authentication methods with
Microsoft Graph - in particular Managed Identity.

If `clientId` and clientSecret` are configured, the plugin will continue to behave
as before using those credentials.  If those values are ommitted, then
`DefaultAzureCredential` is choose the best method of authentication.

This is particularly helpful locally as if the Azure CLI or Azure Powershell are
installed, Azure Identity will use those to authenticate, without the developer
needing to configure anything.  In those cases, the following config will be
sufficient for a working graph plugin.

```yaml
catalog:
  provider:
    microsoftGraphOrg:
      default:
        tenantId: ${AZURE_TENANT_ID}
        user:
          filter: accountEnabled eq true and userType eq 'member'
        group:
          filter: securityEnabled eq false and mailEnabled eq true and groupTypes/any(c:c+eq+'Unified')
```

Signed-off-by: Alex Crome <afscrome@users.noreply.github.com>